LayoutTests/http/tests/security/video-cross-origin-caching-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.

This test passes if you do not see a CORS error.

EVENT(playing)
EVENT(playing)
END OF TEST
